refactor: move SynthesisModels to finetuner.model
